Refactor search bar
