+++ assets/components/Header.js
@@ -3,22 +3,11 @@ import { Link } from 'react-router-dom';
 import jwt_decode from 'jwt-decode'; //npm install jwt-decode
 
 import ButtonUp from './UI components/ButtonUp';
+import SearchBar from './SearchBar';
 
 const Header = ({ books }) => {
-  const [inputField, setInputField] = useState('');
   const [user, setUser] = useState({});
   const [isLoggingOut, setIsLoggingOut] = useState(false);
-
-  const handleSubmit = e => {
-    e.preventDefault();
-  };
-
-  // Input from users
-  const handleChange = e => {
-    //in case someone has put spaces in the input and format it to look nicer in the URL using +
-    const inputFormat = e.target.value.trim().replaceAll(' ', '+');
-    setInputField(inputFormat);
-  };
 
   {
     /*......... GOOGLE LOG IN AND SIGN OUT...........*/
@@ -60,39 +49,28 @@ const Header = ({ books }) => {
           className='navbar-collapse justify-content-between'
           id='navbarSupportedContent'
         >
+          {/*................. Navigation links .................*/}
+          {/* Home */}
           <h1 className='ml-5'>
             <Link to='/' className='text-white'>
               Book Nook
             </Link>
           </h1>
-
-          {/* Send the inputField as :keyword to SearchPage to display it in URL and useParams().keyword to fetch query*/}
-          <form onSubmit={handleSubmit}>
-            <input
-              type='text'
-              onChange={handleChange}
-              defaultValue={inputField}
-              placeholder='Search...'
-            />
-            <button type='submit' className='btn btn-primary'>
-              <Link to={`/search/${inputField}`} className='text-light'>
-                Submit
-              </Link>
-            </button>
-          </form>
-
+          <SearchBar />
+          {/* Home */}
           <ul className='nav'>
             <li className='nav-item'>
               <Link to='/' className='nav-link active text-white'>
                 Home
               </Link>
             </li>
+            {/* Contacts */}
             <li className='nav-item'>
-              {/*......... Scroll down to contacts on any page...........*/}
               <a href='#contact' className='nav-link active text-white'>
                 Contacts
               </a>
             </li>
+            {/* Shopping cart */}
             <li className='nav-item'>
               <Link to='/shopping-cart' className='nav-link active'>
                 <span className='text-white'>{books.length ?? 0 + ' '}</span>
@@ -101,7 +79,7 @@ const Header = ({ books }) => {
                 </span>
               </Link>
             </li>
-
+            {/*............ Google Sign in button ...........*/}
             {/*......... Sign in ...........*/}
             <li id='signInDiv'></li>
             {isLoggingOut && Object.keys(user).length != 0 && (
